+++ src/components/_classes/component/Component.js
@@ -1232,6 +1232,16 @@ export default class Component extends Element {
     return data;
   }
 
+  itemValueForHTMLMode(value) {
+    if (Array.isArray(value)) {
+      const values = value.map(item => Array.isArray(item) ? this.itemValueForHTMLMode(item) : this.itemValue(item));
+
+      return values.join(', ');
+    }
+
+    return this.itemValue(value);
+  }
+
   createModal(element, attr) {
     const dialog = this.ce('div', attr || {});
     this.setContent(dialog, this.renderTemplate('dialog'));
@@ -1787,7 +1797,7 @@ export default class Component extends Element {
     }
 
     // If we are supposed to validate on blur, then don't trigger validation yet.
-    if (this.component.validateOn === 'blur' && !this.errors.length) {
+    if (this.component.validateOn === 'blur' && !this.errors.length && !flags.fromSubmission) {
       flags.noValidate = true;
     }
 
@@ -2527,7 +2537,7 @@ export default class Component extends Element {
     if (!this.builderMode && !this.options.preview && !this.isEmpty(this.defaultValue) && !flags.noValidate) {
       return this.checkComponentValidity(data, true, row);
     }
-    return flags.noValidate ? true : this.checkComponentValidity(data, false, row);
+    return flags.noValidate ? true : this.checkComponentValidity(data, !!flags.fromSubmission, row, flags);
   }
 
   get validationValue() {